requirements.txt

easydict==1.13
einops==0.8.0
numpy==1.24.3
pandas==1.5.3

PyYAML==6.0.2
reformer_pytorch==1.4.4
scikit_learn==1.2.2

sympy==1.12
torch==2.1.1
tqdm==4.66.4
utils==1.0.2
